Implement header styles: banner, userbar, tabs, logos, subreddit list.
Banners (images/banner*) are listed in the .banner file under the stylesheet to which they apply. Each banner object lists the image in which it can be found (multiple banners of the same theme are stored in the same file), the position within that file where they are located, and their height. When Raking, Ruby will replace objects in variables.sass.erb with the values scraped from .banner.

At present, the desired banner must be specified in Rakefile. This will be addressed in the future.